requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
Django==2.2
django-advanced-filters==1.3.0
django-braces==1.14.0
django-cors-headers==3.8.0
django-cors-middleware==1.3.1
django-crispy-forms==1.12.0
django-extensions==2.1.5
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.9.1
idna==3.2
packaging==21.0
Pillow==8.3.1
pip-review==1.1.0
psycopg2==2.9.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.12.0
sqlparse==0.4.1
urllib3==1.26.6
